refactor(SimulationAgent): move classes
