+++ src/main/scala/ulisse/applications/useCases/SimulationService.scala
@@ -6,7 +6,7 @@ import ulisse.applications.{AppState, SimulationState}
 import ulisse.applications.ports.{SimulationPorts, UtilityPorts}
 import ulisse.entities.Coordinate
 import ulisse.entities.route.RouteEnvironmentElement
-import ulisse.entities.simulation.Agents.SimulationAgent
+import ulisse.entities.simulation.SimulationAgent
 import ulisse.entities.simulation.Environments.SimulationEnvironment
 import ulisse.entities.simulation.Simulations.EngineState
 import ulisse.entities.station.Station
