+++ src/main/scala/ulisse/entities/simulation/SimulationAgent.scala
@@ -4,6 +4,5 @@ import ulisse.entities.simulation.Environments.SimulationEnvironment
 import ulisse.entities.simulation.Simulations.Actions.SimulationAction
 import ulisse.entities.train.Trains.Train
 
-object Agents:
-  trait SimulationAgent:
-    def doStep(dt: Int, simulationEnvironment: SimulationEnvironment): SimulationAction
+trait SimulationAgent:
+  def doStep(dt: Int, simulationEnvironment: SimulationEnvironment): Option[SimulationAction]
