+++ src/main/scala/ulisse/entities/station/StationEnvironmentElement.scala
@@ -1,0 +1,5 @@
+package ulisse.entities.station
+
+class StationEnvironmentElement {
+
+}
