+++ src/main/scala/ulisse/entities/train/TrainAgent.scala
@@ -1,9 +1,10 @@
 package ulisse.entities.train
 
-import ulisse.entities.simulation.Agents.SimulationAgent
+import ulisse.entities.simulation.SimulationAgent
 import ulisse.entities.train.Trains.Train
 
 trait TrainAgent extends Train with SimulationAgent:
-  def travelDistance: Double
-  def travelDistance_=(newTravelDistance: Double): TrainAgent
-  def updateTravelDistance(distanceDelta: Double): TrainAgent
+  def distanceTravelled: Double
+  def distanceTravelled_=(newTravelDistance: Double): TrainAgent
+  def resetDistanceTravelled(): TrainAgent                       = distanceTravelled = 0
+  def updateDistanceTravelled(distanceDelta: Double): TrainAgent = distanceTravelled += distanceDelta
